move is element selected helper
