+++ lib/shared/components/dnd/droppable/index.js
@@ -1,6 +1,6 @@
 import * as dndActions from 'actions/dnd';
 import Component from 'components/component';
-import isElementSelected from 'helpers/is-element-selected';
+import isElementSelected from 'helpers/page-builder/is-element-selected';
 import React, {PropTypes} from 'react';
 import {bindActionCreators} from 'redux';
 import {connect} from 'react-redux';
