+++ lib/shared/elements/dynamic-list/container.js
@@ -1,5 +1,5 @@
 import Component from 'components/component';
-import isElementSelected from 'helpers/is-element-selected';
+import isElementSelected from 'helpers/page-builder/is-element-selected';
 import React, {PropTypes} from 'react';
 import {dataConnect} from 'relate-js';
