+++ lib/shared/helpers/page-builder/is-element-selected.js
