+++ lib/shared/screens/admin/shared/components/page-builder-menu/tabs/layers/element/index.js
@@ -6,7 +6,7 @@ import calculateElement from 'helpers/element/calculate';
 import get from 'lodash/get';
 import getElement from 'helpers/page-builder/get-element';
 import getLinksValues from 'helpers/element/get-links-values';
-import isElementSelected from 'helpers/is-element-selected';
+import isElementSelected from 'helpers/page-builder/is-element-selected';
 import React, {PropTypes} from 'react';
 import {connect} from 'react-redux';
