+++ lib/shared/screens/admin/shared/components/page-builder/canvas/page-element/element/index.js
@@ -1,5 +1,5 @@
 import Component from 'components/component';
-import isElementSelected from 'helpers/is-element-selected';
+import isElementSelected from 'helpers/page-builder/is-element-selected';
 import React, {PropTypes} from 'react';
 import {bindActionCreators} from 'redux';
 import {connect} from 'react-redux';
